Try rewriting asserts regardless of text of line
The code previously was looking for the word "assert" on the line of
code where an `AssertionError` was raised before being willing to try
rewriting asserts and trying the test again.  This means that multi-line
assert statements were never retried, because the line of text was then
the last line of the statement, not the first.  So, now, we just always
optimistically try rewriting as long as the error was `AssertionError`.

While dedenting this code, I also went ahead and brought the width under
80 columns.  Why was I writing such long lines?