Fix line number, for Python 3.10 compatibility
In Python previous to 3.10, the traceback for:
```
    with assert_raises(ValueError, 'one message'):
        raise ValueError('another message')
```
would, if the `assert_raises()` context manager raised an exception,
leave the line number on the second line.  So any complaint we raised
about the exception message would get printed right next to the `raise`
that is being complained about.

But in Python 3.10 there are two problems.  First, the `frame.f_lineno`
line number we were using steps backward to point at the `with` line,
not the `raise` line, as Python enters the context manager `__exit__()`
method; so we need to switch to `tb.tb_lineno`, which it looks like is
what the official traceback code has been using all along.  Second, the
traceback created from raising our own exception in `__exit__()` has
both line numbers pointing at the `with` line, not the `raise()` line.
Solution: store the original traceback on a custom exception.