Fancy assert for “in” operator in Python 3.9, 3.10
This fairly thorough cleanup of the `assertion` module switches the
bytecode regular expression to matching the entire comparison
instruction instead of hard-coding the opcode ‘COMPARE_OP’, because
operators like ‘in’ and ‘is’ now have their own opcodes.

To simplify the code and testing, we hereby abandon Python 2.6, which
was already not listed in our `setup.py`.  Users can hopefully upgrade
to Python 2.7 painlessly.